Ensure a unique storage is used for the VM
Create a storage account mainly to avoid the warning (and issue) of
New-AzVM defaulting to whatever storage it finds, even in other
resource groups.

A potential extra benefit is to use it for the Skyscraper cache, but I
haven't explored this as my config scripts assume the drive mounted to a
specific path (matching RetroPie).