rpi: Fix storage mounting issue on VM
Running "sudo mount -a" in the VM mount-share.sh would lead to a
"permission denied". It was caused by the storage account key having "="
characters stripped out during the creation of the VM.

Now storing the same environment variables on the rpi as on the vm, to
have something to check against.