rpi: Change mount to attempt at mount
It can't run in CI/CD with Docker on CircleCI because that requires a
"machine" executor to run containers in privileged mode, which doesn't
support custom images such as ours.

This mounting step has had issues on RPi previously as well. Likely due
access rights issues, but can also be network issues. As it's a sensitive
part of the setup script, which can be recovered later with a reboot, it's
better to not exit with an error.

Note: The following steps with copying the ROMs automatically is likely
to fail and should be moved to a manual script.