docker: Create docker image with Skyscraper
Installing Skyscraper on a new VM is very slow. By having an image with it
pre-installed it will greatly speed up the install retro-cloud for users,
and the development of retro-cloud.

The downside is maintaining the image as Skyscraper releases new versions.
Depending on how long upgrading takes, it could still be a step during
setup.