docker: Remove retro-cloud source
The development pipeline is different from when this was added. The
download-and-run.sh can take any branch or commit, which makes it easier
to run any test version. Having old source in the docker image only
interferes.

Add a dockerignore file that excludes everything from the daemon
context, since we're no longer going to include any files in the image.
Also makes building slightly faster.