ci: Add a copy ROM step to better verify the scraper
Containers are not running with privileges so the ROM isn't copied to
the Azure File Share, so that part can't be verified. It will have to be
enough with a local test for now.

Without sudo it can't write to the shared folder. Probably related to
the failed mount (although successful mount has the same error, it does
download the file).