vm: Make scraper script chatty
Silent install is nice and should be an option in the future, but during
development it's just time consuming to not see what went wrong.
Skyscraper is very chatty, and my perception is that the output slows
down the processing, but wll enable it for now anyway to make sure it's
running properly.