docker: Clean up unnecessary files at the end
Clean up the files at the end to shorten build time as package sources
don't have to be repopulated, and the final image is truly cleaned.

Also clean some more based on some references. It's less than 50MB,
which is nothing compared to a 1.27GB image, but it's a good practice.