docker: Run RetroPie-Setup as RPi 3 on arm32v7
Set platform flag as RPi 3 for RetroPie-Setup when running on a arm32v7
processor, assuming the image is being built on a real RPi 3. Unsure
what the flag does but requires multiple packages from the RaspberryPi
foundation repo, which needs sources and their certificates to be added
to apt.

About the platform flag:
* Noticed it in lasery/retropie Dockerfile:
  https://github.com/laseryuan/docker-apps/blob/db3c154ebebb20613125b1669483450725c79a63/retropie/Dockerfile.templ#L37      
* Searched through RetroPie-Setup and found this:
  https://github.com/RetroPie/RetroPie-Setup/blob/729797916a1259165f03b24fd46672ddd9ba5c8e/scriptmodules/system.sh#L312-L325

About the architecture check:
Note the single-equal. Double-equal is specific to bash so it didn't
work in Docker.