docker: Lockdown RetroPie-Setup to 4.5.16
There was a new release, 4.5.17, which was being used automatically and
the setup was silently failing, but it only had one emulator so local
development scenarios were suddenly not working.

It turned out to be unrelated to the release but still decided to start
locking down the version to avoid surprises. Especially while working
on the arm32 Docker image as it makes testing easier to isolate issues.