docker: Workaround for debconf warning about Dialog
Configure Debian frontent to be non-interactive. Fixes this warning:
"debconf: unable to initialize frontend: Dialog"

Note: This is not recommended as it affects the built image, so all
users of the image has this setting although it wouldn't make sense for
them. The preferred approach is to use `ARG DEBIAN_FRONTEND=noninteractive`
but this doesn't seem to work on Ubuntu (attempted in `docker-debconf-dialog`).
Will revisit in `docker-multiarch`.