test(docker): Verify file endings
Sometimes when building the image on Windows the user bash profile files
have Carriage Return (`\n`) in the line endings, which confuses the bash
shell so it doesn't load the profile correctly and gives a warning. The
RPi files in docker/rpi should have LF but sometimes git checks it out
as CRLF, despite `.gitattributes`, and the COPY step in the Dockerfile
doesn't convert.