docker: Fix error "fuse: failed to open /dev/fuse: Permission denied"
Happens when mounting the VM share in mount-vm-share.sh.

The issue has two parts:
1. "pi" does not have write access to `/dev/fuse`, not even with sudo,
   so mount-vm-share.sh fails to mount the drive.
2. `/dev/fuse` does not exist until runtime so it's not accessible when
   building the Docker image.

The easiset solution is to give back root-membership to "pi". It was
removed in the switch from `useradd` to `adduser` in PR #8 (67a90676).

Not the same, but there might be something to learn from Docker's
requirement on root access for mounting volumes:
https://github.com/moby/moby/issues/2259

Note: The "pi" user on a real RetroPie installation on a RPi is _not_
member of root. So this solution should be removed one day.

Cleaned up Dockerfile. The instructions will be added to Readme.