rpi: Handle a freshly installed RetroPie
After installing RetroPie and it rebooting it goes into
EmulationStation, but if the user exits without configuring the
controllers and runs Retro-Cloud setup it exited when trying to back up
the ~/.emulationstation/downloaded_media.

I'm assuming it will be much more common to install Retro-Cloud on a RPi
that's been used to play games on before, so the Docker image is not
changed. Just clarified what the fake file structure is about.