rpi: Move env vars to a separate `~/.retro-cloud.dev`
Moving all the environmental variables from .bashrc makes it
easier to copy Azure details to:
* VM
* Docker host (my dev machine)
* Docker container (to not have to setup Azure again)
* CI artifacts

Also keeps the users .bashrc file cleaner.

It can also allow the .env file to become a "template" to be filled in.
Making it easier to see what is expected, and allow for the scripts
to fill it in as it runs. Which would make it much easier to recover
from failures (today it's almost impossible as the script must complete
without errors to save any Azure details).

Save it in the users HOME path (and not in eg. /etc/profile.d/) because
authenticating to the created resources requires the users SSH key, so
it's not re-usable by different users.