rpi: Fetch Azure File Share key when testing
Removing the use of a stored key to increase security. Especially
avoids it showing up in public build logs (currently not a concern as
those environments gets torn down immediately).

By decreasing the need to store Azure resource information locally it
will eventually allow for a develop Azure infrastructure to be
recreated and still be able to connect to it.