dev: Use host's SSH, and EmulationStation configs
Read-only volumes don't always work out because the application needs
to write to those directories, but we don't want those changes to leak
out of the container. So we create the container first and then copy
over files to it.

Note: `docker-run.sh` shouldn't include the host's SSH or other pieces
that do not resemble a user's RetroPie.