ci: Use the "commit" workflow to trigger CircleCI builds
It's a preview feature but it works. Keeping the builds grouped by
workflow makes it much easier to get an overview of progress in CircleCI's
web UI, and more likely to queue builds in a way that don't run more
scriptValidation jobs than Azure allows for public IP's (max 10/region).

Example:
https://circleci.com/docs/api/#trigger-a-new-build-by-project-preview