ci: Disable CircleCI
Kept one so the YAML doesn't break. I had a hard time getting CircleCI to work in the beginning due to a broken config on master branch that would somehow affect CircleCI so it wouldn't detect my fixed config in another branch. So I don't want to mess with that again.