docker: Add rpi (arm32v7) docker compose override
Possible to run on a RPi with:
```
$ docker-compose -f docker-compose.yml -f docker-compose.arm32v7.yml run --rm rpi bash
```

The RPi image should already have the "pi" user as part of the same groups
as a real RPi, but it only really needed when running the container on a
real RPi. With the limitation on Docker Compose and groupadd I added the
two main groups in the Dockerfile.