docker: Create a containerized Skyscraper image
Installing Skyscraper on a new VM is very slow. By having a tiny image
with it as a executable, similar to `shellcheck` that we use for linting
Bash scripts. It will greatly speed up the install of retro-cloud on the
VM for users, and for developers.

This image is tiny: 90MB
Compared to a full Unbuntu image with Skyscraper installed: 598MB

The downside is maintaining the image as Skyscraper releases new versions.
Depending on how long upgrading takes, it could still be a step during
setup.

About the Dockerfile:
The multi-stage Dockerfile will create a baseline image that matches the
runtime. Then add the build tools and build Skyscraper. Finally, copy the
executable back over to a runtime image.

The container will now run Skyscraper and exit. This should make it
easier to use as a dockerized app, like "shellcheck". Which should work
well in a swarm-like setup, for spinning up multiple Skyscrapers in Azure
and then shutting them down once the scraping is complete.

Also tried
```
apt-get autoclean && apt-get clean
```
but it didn't change the size.