vm: Install Docker on VM during install-skyscraper
The VM created doesn't include Docker so it needs to be installed to run
Skyscraper as a containerized app. Using the convenience script to install
Docker is not recommended, but it's still closer to Docker supporting it than
the docker.io image that's maintained by Ubuntu that was behaving the same.
https://docs.docker.com/engine/install/ubuntu/#install-using-the-convenience-script

* Add user to docker group to avoid sudo
* Log in to a new group instead of rebooting (`newgrp docker`)

This works when running the scripts on a VM created from a local
container, but on CircleCI it's failing on `docker pull`:
```
docker: Got permission denied while trying to connect to the Docker daemon socket at unix:///var/run/docker.sock: Post http://%2Fvar%2Frun%2Fdocker.sock/v1.40/containers/create: dial unix /var/run/docker.sock: connect: permission denied.
```

Note: The uninstall script is taken from Docker docs but is untested.