dev: Add build scripts for the containerized Skyscraper
Print build commands before running them, but do it after including the
secrets so they are not printed.

Improved logging compared to other Docker build scripts as it includes the
steps, making both terminal and log file more readable.

Scripts are in virtual-machine/docker/ to not mix them with the VM files
meant for the user.