ci: Add Travis-CI
The main reason is because it allows running docker containers in
privileged mode. That allows the VM to be mounted on the containerized
RPi. That mount is the connection between the user's RPi and
Retro-Cloud, so I needed a way to have that connection tested
continously to confidentely build new features. CircleCI doesn't
privileged containers, so I will be moving away from it eventually.